+++ src/List.js
@@ -1,18 +1,19 @@
-import { enforceFunction, isArray, methodFunction } from 'type-enforcer';
-
-const defaultSorter = (a, b) => a === b ? 0 : a < b ? -1 : 1;
-const numericSortAsc = (a, b) => a - b;
-const numericSortDesc = (a, b) => b - a;
+import defaultCompare from 'default-compare';
+import { isArray, methodFunction } from 'type-enforcer';
 
 const sorters = Object.freeze({
-	default: defaultSorter,
+	default: defaultCompare,
 	string: {
 		asc: (a, b) => a.localeCompare(b),
 		desc: (a, b) => b.localeCompare(a)
 	},
 	number: {
-		asc: numericSortAsc,
-		desc: numericSortDesc
+		asc: (a, b) => a - b,
+		desc: (a, b) => b - a
+	},
+	id: {
+		asc: (a, b) => defaultCompare(a, b, 'id'),
+		desc: (a, b) => defaultCompare(b, a, 'id')
 	}
 });
 
@@ -21,8 +22,6 @@ export const sortedIndexOf = (array, item, sorter, isInsert = false, isLast = fa
 	let low = 0;
 	let mid;
 	let diff;
-
-	sorter = enforceFunction(sorter, defaultSorter);
 
 	while (low < high) {
 		mid = high + low >>> 1;
@@ -45,16 +44,17 @@ export const sortedIndexOf = (array, item, sorter, isInsert = false, isLast = fa
 const ARRAY = Symbol();
 
 /**
+ * @class List
+ * @summary
+ *
+ * ``` javascript
+ * import { List } from 'hord';
+ * ```
+ * @description
  * Always sorted array.
  *
  * List maintains a sorted state internally, but doesn't observe changes to it's contents, so items manipulated externally can cause problems. If you must do this, the .sort() method is provided to resort the list.
  *
- * ## Usage
- * ``` javascript
- * import { List } from 'hord';
- * ```
- *
- * @class List
  *
  * @arg {Array} [values]
  */
@@ -68,11 +68,11 @@ export default class List {
 	 *
 	 * @memberOf List
 	 * @instance
+	 * @chainable
 	 */
 	sort() {
 		if (this[ARRAY].length) {
-			const sorter = this.sorter();
-			this[ARRAY].sort(sorter === sorters.default ? undefined : sorter);
+			this[ARRAY].sort(this.sorter());
 		}
 		return this;
 	}
@@ -82,10 +82,10 @@ export default class List {
 	 *
 	 * @memberOf List
 	 * @instance
+	 * @chainable
 	 *
 	 * @arg {*} item - Item is inserted into the list such that the items are still sorted.
 	 *
-	 * @returns {this}
 	 */
 	add(item) {
 		this[ARRAY].splice(sortedIndexOf(this[ARRAY], item, this.sorter(), true) + 1, 0, item);
@@ -94,14 +94,61 @@ export default class List {
 	}
 
 	/**
-	 * Merges one or more arrays with the list.
+	 * Add an item to the list if it isn't already included. Uses binary search.
+	 *
+	 * @memberOf List
+	 * @instance
+	 * @chainable
+	 *
+	 * @arg {*} item - Item is inserted into the list such that the items are still sorted.
+	 */
+	addUnique(item) {
+		if (this[ARRAY].length === 0) {
+			this[ARRAY][0] = item;
+		}
+		else {
+			const sorter = this.sorter();
+			let index = sortedIndexOf(this[ARRAY], item, sorter, true);
+
+			if (index === -1 || sorter(this[ARRAY][index], item) !== 0) {
+				this[ARRAY].splice((index || -1) + 1, 0, item);
+			}
+		}
+
+		return this;
+	}
+
+	/**
+	 * Get a new List of the unique (as determined by the sorter) values in this List.
 	 *
 	 * @memberOf List
 	 * @instance
 	 *
-	 * @arg {*} values
+	 * @returns {List}
+	 */
+	unique() {
+		const sorter = this.sorter();
+		const output = [];
+		let previous;
+
+		this.forEach((item, index) => {
+			if (index === 0 || sorter(previous, item)) {
+				output.push(item);
+				previous = item;
+			}
+		});
+
+		return new List().sorter(sorter).values(output);
+	}
+
+	/**
+	 * Merges one or more arrays with the list.
 	 *
-	 * @returns {this}
+	 * @memberOf List
+	 * @instance
+	 * @chainable
+	 *
+	 * @arg {*} values
 	 */
 	concat(...args) {
 		const self = this;
@@ -114,13 +161,25 @@ export default class List {
 	 *
 	 * @memberOf List
 	 * @instance
+	 * @chainable
 	 *
 	 * @arg {*} item - Uses the sorter function to determine equality.
-	 *
-	 * @returns {this}
 	 */
 	discard(item) {
 		this[ARRAY].splice(sortedIndexOf(this[ARRAY], item, this.sorter(), true), 1);
+
+		return this;
+	}
+
+	/**
+	 * Discard all items from the list.
+	 *
+	 * @memberOf List
+	 * @instance
+	 * @chainable
+	 */
+	discardAll() {
+		this[ARRAY].length = 0;
 
 		return this;
 	}
@@ -130,6 +189,7 @@ export default class List {
 	 *
 	 * @memberOf List
 	 * @instance
+	 * @chainable
 	 *
 	 * @arg {Array} [values] - If provided, replaces any previous values with these, otherwise return the current values.
 	 *
@@ -152,7 +212,7 @@ export default class List {
 	 *
 	 * @arg {*} item - Uses the sorter function to determine equality.
 	 *
-	 * @returns {Number} - The index of the item or -1
+	 * @returns {Number} The index of the item or -1
 	 */
 	indexOf(item) {
 		return sortedIndexOf(this[ARRAY], item, this.sorter());
@@ -166,7 +226,7 @@ export default class List {
 	 *
 	 * @arg {*} item - Uses the sorter function to determine equality.
 	 *
-	 * @returns {Number} - The index of the item or -1
+	 * @returns {Number} The index of the item or -1
 	 */
 	lastIndexOf(item) {
 		return sortedIndexOf(this[ARRAY], item, this.sorter(), false, true);
@@ -194,7 +254,7 @@ export default class List {
 	 *
 	 * @arg {*} item - Uses the sorter function to determine equality.
 	 *
-	 * @returns {*} - The item or undefined
+	 * @returns {*} The item or undefined
 	 */
 	find(item) {
 		return this[ARRAY][this.indexOf(item)];
@@ -208,7 +268,7 @@ export default class List {
 	 *
 	 * @arg {*} item - Uses the sorter function to determine equality.
 	 *
-	 * @returns {*} - The item or undefined
+	 * @returns {*} The item or undefined
 	 */
 	findLast(item) {
 		return this[ARRAY][this.lastIndexOf(item)];
@@ -222,10 +282,12 @@ export default class List {
 	 *
 	 * @arg {*} item - Uses the sorter function to determine equality.
 	 *
-	 * @returns {Array} - An array of items or undefined
+	 * @returns {List} A list of items
 	 */
 	findAll(item) {
-		return this[ARRAY].slice(this.indexOf(item), this.lastIndexOf(item) + 1);
+		return new List()
+			.sorter(this.sorter())
+			.values(this[ARRAY].slice(this.indexOf(item), this.lastIndexOf(item) + 1));
 	}
 
 	/**
@@ -236,7 +298,7 @@ export default class List {
 	 *
 	 * @arg {*} item - Uses the sorter function to determine equality.
 	 *
-	 * @returns {Number} - The index of the item or -1
+	 * @returns {Number} The index of the item or -1
 	 */
 	findIndex(item) {
 		return this.indexOf(item);
@@ -250,7 +312,7 @@ export default class List {
 	 *
 	 * @arg {*} item - Uses the sorter function to determine equality.
 	 *
-	 * @returns {Number} - The index of the item or -1
+	 * @returns {Number} The index of the item or -1
 	 */
 	findLastIndex(item) {
 		return this.lastIndexOf(item);
@@ -285,6 +347,7 @@ export default class List {
 	 *
 	 * @memberOf List
 	 * @instance
+	 * @readonly
 	 *
 	 * @returns {Number}
 	 */
@@ -398,18 +461,6 @@ List.prototype.sorter = methodFunction({
  * @returns {Boolean}
  */
 /**
- * See [Array.prototype.filter()](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/filter)
- *
- * @method filter
- * @memberOf List
- * @instance
- *
- * @arg {Function} callback
- * @arg {Object} [thisArg]
- *
- * @returns {Array}
- */
-/**
  * See [Array.prototype.forEach()](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/forEach)
  *
  * @method forEach
@@ -481,18 +532,6 @@ List.prototype.sorter = methodFunction({
  * @returns {*}
  */
 /**
- * See [Array.prototype.slice()](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/slice)
- *
- * @method slice
- * @memberOf List
- * @instance
- *
- * @arg {Number} [begin=0]
- * @arg {Number} [end=array.length]
- *
- * @returns {Array}
- */
-/**
  * See [Array.prototype.some()](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/some)
  *
  * @method some
@@ -506,17 +545,51 @@ List.prototype.sorter = methodFunction({
  */
 [
 	'every',
-	'filter',
 	'forEach',
 	'toLocaleString',
 	'join',
 	'map',
 	'reduce',
 	'reduceRight',
-	'slice',
 	'some'
 ].forEach((key) => {
 	List.prototype[key] = function(...args) {
 		return this[ARRAY][key](...args);
+	};
+});
+
+/**
+ * See [Array.prototype.filter()](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/filter)
+ *
+ * @method filter
+ * @memberOf List
+ * @instance
+ *
+ * @arg {Function} callback
+ * @arg {Object} [thisArg]
+ *
+ * @returns {List}
+ */
+/**
+ * See [Array.prototype.slice()](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/slice)
+ *
+ * @method slice
+ * @memberOf List
+ * @instance
+ *
+ * @arg {Number} [begin=0]
+ * @arg {Number} [end=array.length]
+ *
+ * @returns {List}
+ */
+[
+	'filter',
+	'slice'
+].forEach((key) => {
+	List.prototype[key] = function(...args) {
+		const newList = new List()
+			.sorter(this.sorter());
+		newList[ARRAY] = this[ARRAY][key](...args);
+		return newList;
 	};
 });